requirements.txt
langchain==0.1.0
langchain-community==0.0.20
python-dotenv==1.0.0
pandas==2.0.3